+++ Tests/BenchmarkPerformance/Configurations/ThroughputTestConfiguration.cs
@@ -4,6 +4,7 @@ using BenchmarkDotNet.Diagnosers;
 using BenchmarkDotNet.Engines;
 using BenchmarkDotNet.Exporters;
 using BenchmarkDotNet.Exporters.Csv;
+using BenchmarkDotNet.Exporters.Json;
 using BenchmarkDotNet.Filters;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Toolchains.InProcess.Emit;
@@ -16,54 +17,60 @@ internal class ThroughputTestConfiguration : ManualConfig
 {
     public ThroughputTestConfiguration(string name)
     {
-        if (name != null && name.Equals("dry", StringComparison.InvariantCultureIgnoreCase))
+        name ??= string.Empty;
+
+        var job = Job.Default
+            .WithIterationCount(10)
+            .WithUnrollFactor(1)
+            .WithInvocationCount(10)
+            .WithWarmupCount(5)
+            .WithStrategy(RunStrategy.Throughput);
+
+        if (name.Equals("dry", StringComparison.InvariantCultureIgnoreCase))
         {
-            AddJob(Job.Dry
-                .WithStrategy(RunStrategy.Throughput)
-                .WithArguments([
-                    new MsBuildArgument("/p:Platform=" + BenchmarkHelpers.GetPlatformString())
-                ]));
+            job = Job.Dry.WithStrategy(RunStrategy.Throughput);
         }
-        else if (name != null && name.Equals("short", StringComparison.InvariantCultureIgnoreCase))
+        else if (name.Equals("short", StringComparison.InvariantCultureIgnoreCase))
         {
-            var job = Job.ShortRun
+            job = Job.ShortRun
                 .WithUnrollFactor(1)
                 .WithStrategy(RunStrategy.Throughput);
-
-            // Due to the error building the external process, we need to use the in-process emit toolchain.
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                job = job.WithToolchain(new InProcessEmitToolchain(TimeSpan.FromMinutes(30), true));
-            }
-
-            AddJob(job);
-
-            // Does not run JSON tests in this configuration.
-            AddFilter(new NameFilter(n => !n.Contains("JSON", StringComparison.CurrentCultureIgnoreCase)));
         }
-        else
+
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         {
-            AddJob(Job.Default
-                .WithIterationCount(10)
-                .WithUnrollFactor(1)
-                .WithInvocationCount(10)
-                .WithWarmupCount(5)
-                .WithStrategy(RunStrategy.Throughput)
-                .WithArguments([
-                    new MsBuildArgument("/p:Platform=" + BenchmarkHelpers.GetPlatformString())
-                ]));
+            // Due to the error building the external process in the github-hosted runners,
+            // we need to use the in-process emit toolchain
+            job = job.WithToolchain(new InProcessEmitToolchain(TimeSpan.FromMinutes(60), true));
         }
+
+        AddJob(job);
 
         // Cannot be run without a valid RTI Connext license.
         AddFilter(new NameFilter(n => !n.Contains("RTI", StringComparison.CurrentCultureIgnoreCase)));
 
+        // Does not run JSON tests
+        AddFilter(new NameFilter(n => !n.Contains("JSON", StringComparison.CurrentCultureIgnoreCase)));
+
+        // Diagnosers
+        AddDiagnoser(MemoryDiagnoser.Default);
+
+        // Columns
         AddColumnProvider(DefaultConfig.Instance.GetColumnProviders().ToArray());
+        HideColumns("Gen0", "Gen1", "Gen2", "Alloc Ratio");
         AddColumn(new ThroughputPerSecondColumn());
+
+        // Loggers
         AddLogger(DefaultConfig.Instance.GetLoggers().ToArray());
+
+        // Exporters
         AddExporter(PlainExporter.Default);
         AddExporter(CsvExporter.Default);
         AddExporter(MarkdownExporter.Default);
-        AddDiagnoser(MemoryDiagnoser.Default);
+        AddExporter(JsonExporter.FullCompressed);
+        AddExporter(RPlotExporter.Default);
+
+        // Increase the build timeout to 30 minutes
         WithBuildTimeout(TimeSpan.FromMinutes(30));
     }
 }